Try to fix test suite on Travis CI
What makes this a bit tricky is that the test suite passes for me
locally, this is due to the fact that the Ubuntu mirrors URL
reports different contents based on where it's requested from.

However I think I understand the problem: The test suite was relying on
string matching against URLs which isn't guaranteed to work and has
failed similarly in the past.

My intention with this change is to arrive at a more robust (albeit
slower) solution to the problem: The test suite now tries to fetch
a file at a stable location on each mirror.